backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.9
websockets==13.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4
openai==1.51.0
whisper-timestamped==1.15.0
torch==2.4.1
torchaudio==2.4.1
requests==2.32.3
httpx==0.27.2